requirements.txt
scrapy-splash
scrapyjs
